pyproject.toml
[project]
name = "croc"
version = "0.1.0"
description = "A Rust-inspired, Typer-powered CLI for reliably managing project documentation."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[project.scripts]
croc = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["croc"]
